Implement new structure shape, rename to blender
