+++ src/main/java/supersymmetry/common/metatileentities/multi/electric/MetaTileEntityBlender.java
@@ -1,0 +1,104 @@
+package supersymmetry.common.metatileentities.multi.electric;
+
+import gregtech.api.capability.impl.MultiblockRecipeLogic;
+import gregtech.api.metatileentity.MetaTileEntity;
+import gregtech.api.metatileentity.interfaces.IGregTechTileEntity;
+import gregtech.api.metatileentity.multiblock.IMultiblockPart;
+import gregtech.api.metatileentity.multiblock.RecipeMapMultiblockController;
+import gregtech.api.pattern.BlockPattern;
+import gregtech.api.pattern.FactoryBlockPattern;
+import gregtech.api.pattern.MultiblockShapeInfo;
+import gregtech.api.pattern.TraceabilityPredicate;
+import gregtech.api.unification.material.Materials;
+import gregtech.client.renderer.ICubeRenderer;
+import gregtech.client.renderer.texture.Textures;
+import gregtech.client.utils.TooltipHelper;
+import gregtech.common.ConfigHolder;
+import gregtech.common.blocks.BlockBoilerCasing;
+import gregtech.common.blocks.BlockMetalCasing;
+import gregtech.common.blocks.MetaBlocks;
+import gregtech.common.blocks.BlockTurbineCasing.TurbineCasingType;
+import gregtech.common.metatileentities.MetaTileEntities;
+import gregtech.core.sound.GTSoundEvents;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.resources.I18n;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.SoundEvent;
+import net.minecraft.world.World;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
+import supersymmetry.api.recipes.SuSyRecipeMaps;
+import supersymmetry.common.metatileentities.SuSyMetaTileEntities;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.ArrayList;
+import java.util.List;
+
+public class MetaTileEntityBlender extends RecipeMapMultiblockController {
+
+    public MetaTileEntityBlender(ResourceLocation metaTileEntityId) {
+        super(metaTileEntityId, SuSyRecipeMaps.BLENDER_RECIPES);
+        this.recipeMapWorkable = new MultiblockRecipeLogic(this, true);
+    }
+
+    @Override
+    public MetaTileEntity createMetaTileEntity(IGregTechTileEntity tileEntity) {
+        return new MetaTileEntityBlender(metaTileEntityId);
+    }
+
+    @Override
+    protected BlockPattern createStructurePattern() {
+        TraceabilityPredicate casing = states(getCasingState()).setMinGlobalLimited(34);
+        TraceabilityPredicate abilities = autoAbilities();
+        return FactoryBlockPattern.start()
+                .aisle(" XXX ", " XPX ", " XXX ", "  X  ")
+                .aisle("XXXXX", "X D X", "X   X", "  X  ")
+                .aisle("XXXXX", "PDDDP", "X E X", "XXCXX")
+                .aisle("XXXXX", "X D X", "X   X", "  X  ")
+                .aisle(" XXX ", " XSX ", " XXX ", "  X  ")
+                .where('S', selfPredicate())
+                .where('X', casing.or(abilities))
+                .where('P', states(getPipeCasingState()))
+                .where('D', states(MetaBlocks.METAL_CASING.getState(BlockMetalCasing.MetalCasingType.STAINLESS_CLEAN))) 
+                .where('E', states(MetaBlocks.FRAMES.get(Materials.StainlessSteel).getBlock(Materials.StainlessSteel)))
+                .where('C', states(MetaBlocks.TURBINE_CASING.getState(TurbineCasingType.STAINLESS_STEEL_GEARBOX)))
+                .where(' ', any())
+                .build();
+    }
+
+    @SideOnly(Side.CLIENT)
+    @Override
+    public ICubeRenderer getBaseTexture(IMultiblockPart sourcePart) {
+        return Textures.INERT_PTFE_CASING;
+    }
+
+    protected IBlockState getCasingState() {
+        return MetaBlocks.METAL_CASING.getState(BlockMetalCasing.MetalCasingType.PTFE_INERT_CASING);
+    }
+
+    protected IBlockState getPipeCasingState() {
+        return MetaBlocks.BOILER_CASING.getState(BlockBoilerCasing.BoilerCasingType.POLYTETRAFLUOROETHYLENE_PIPE);
+    }
+
+    @Override
+    public SoundEvent getBreakdownSound() {
+        return GTSoundEvents.BREAKDOWN_ELECTRICAL;
+    }
+
+    @Override
+    public void addInformation(ItemStack stack, @Nullable World player, List<String> tooltip, boolean advanced) {
+        super.addInformation(stack, player, tooltip, advanced);
+        tooltip.add(TooltipHelper.RAINBOW_SLOW + I18n.format("gregtech.machine.perfect_oc"));
+    }
+
+    @SideOnly(Side.CLIENT)
+    @Nonnull
+    @Override
+    protected ICubeRenderer getFrontOverlay() {
+        return Textures.LARGE_CHEMICAL_REACTOR_OVERLAY;
+    }
+
+}
